+++ src/app.rs
@@ -1,12 +1,13 @@
+use heapless::Vec;
 use trussed::interrupt::InterruptFlag;
 
 pub use crate::command::Command;
-pub use crate::types::{AppResult, Error, Message};
+pub use crate::types::{AppResult, Error};
 
 /// trait interface for a CTAPHID application.
 /// The application chooses which commands to register to, and will be called upon
 /// when the commands are received in the CTAPHID layer.  Only one application can be registered to a particular command.
-pub trait App<'interrupt> {
+pub trait App<'interrupt, const N: usize> {
     /// Get access to the app interrupter
     fn interrupt(&self) -> Option<&'interrupt InterruptFlag> {
         None
@@ -19,5 +20,5 @@ pub trait App<'interrupt> {
     /// Application must put response in @message, or decide to return an error.
     ///
     /// The response is pre-cleared.
-    fn call(&mut self, command: Command, request: &Message, response: &mut Message) -> AppResult;
+    fn call(&mut self, command: Command, request: &[u8], response: &mut Vec<u8, N>) -> AppResult;
 }